Fixed sanitisation of source names in makeMetricTable()
Previously, calling makeMetricTable() with a source name containing a
hypen (or possibly other non-alphanumeric characters) would cause the
function to return a malformed dataframe, with two additional columns
with NA for names.

This was ultimately caused by the fact that calling we had a call to
data.frame() with check.names = TRUE (default behaviour), which was
sanitising the column names, some of which are constructed from the
source names. Later code then attempted to manually sanitise the source
names when selecting a column to write to, but the manual sanitisation
only accounted for whitespace.

The solution I've implemented here is to use check.names = FALSE when
calling data.frame(), which preserves the original column names.